Added configurable MapTips, fixed problem with Legend and TOC being in opposite order
A MapTip is a tooltip that appears when you hover over a feature.  The
content of the MapTip is configurable.  The feature is highlighted by a
configurable symbol (SimpleMarkerSymbol, SimpleLineSymbol, or
SimpleFillSymbol).  If any of the attributes defined in the MapTip have
no value, then they will display a configurable string.  There is a
checkbox in the header for turning MapTips on and off.

Thw Legend and TOC had their layers in the opposite order.  The Legend
was correct.  I fixed the TOC.