reset mapClickMode if StreetViewStatus has Zero_Results
If the basemap is not in the spatial reference StreetView needs to convert the coordinates then the mapClickMode does not reset to the default mode. The discussion was originally posted here https://github.com/DavidSpriggs/ConfigurableViewerJSAPI/issues/84#issuecomment-48631765

A followup comment by @tmcgee suggested placing my change outside the If statement.