completed implementation
with this approach, enabling disabling infoTemplates is separated into a separate module and called publishing to either:
'mapInfoTemplates/enable'
'mapInfoTemplates/disable'

This new module, InfoTemplateManager, is setup as a property of the Controller during the startup method.

The InfoTemplateManager will attempt to enable/disable infoTemplates by using the setInfoWindowOnClick method if it exists.  If not, it will use Tim's approach of storing infoTemplate/infoTemplates objects for each layer in the map.