Add topics to Identify Widget
Added the following published topics:

1. identify/execute
2. identify/results
3. identify/error

These are intended to allow more control for widgets like IdentifyPanel.

Added the following subscribed topic:

1. identify/removeLayerInfos

This allows for removing the layer when it is removed from the map such as the DragAndDrop widget. This addition required changes when adding LayerInfos to capture the error handler for later removal.

Added a `showPopup` property so visibility of the map's infoWindow can be managed outside of the Identify widget (such as from the IdentifyPanel widget).